Move detection of runtype
