+++ taca/organise/flowcells.py
@@ -4,9 +4,28 @@ import logging
 import os
 
 from taca.utils.config import CONFIG
+from taca.utils.misc import call_external_command_detached
 
 logger = logging.getLogger(__name__)
 
+def get_flowcell_type(flowcell):
+    """Return flowcell type based on flowcell name"""
+    pass
+
+def instantiate_flowcell(flowcell, project):
+    flowcell_type = get_flowcell_type(flowcell)
+
+    if flowcell_type == "nanopore":
+        return NanoporeFlowcell(flowcell=flowcell, project_id=project)
+    elif flowcell_type == "illumina":
+        return IlluminaFlowcell(flowcell=flowcell, project_id=project)
+    elif flowcell_type == "element":
+        return ElementFlowcell(flowcell=flowcell, project_id=project)
+    else:
+        logger.warning(
+            f"Flowcell type could not be recognised for flowcell {flowcell}, skipping it."
+        )
+        return
 
 class Flowcell:
     """Defines a generic Flowcell"""
@@ -36,7 +55,9 @@ class NanoporeFlowcell(Flowcell):
     def organise_data(self):
         """Tarball data into ONT_TAR"""
         # generate tarball (detached process? how to signal when tarballing is done?)
-        pass
+        # future todo: also organise data in DATA for easier analysis
+        tar_command = f'tar -cf {self.fc_id}'
+        call_external_command_detached(tar_command)
 
 
 class IlluminaFlowcell(Flowcell):
